storage services added

+++ app/services/user_services.py
@@ -1,8 +1,11 @@
+from sqlite3 import IntegrityError
+
 from sqlalchemy.ext.asyncio import AsyncSession
 from app.exceptions.user_exceptions import UserNotFoundError
 from app.models.database import User
 from sqlalchemy import select
-from fastapi import HTTPException
+from fastapi import HTTPException, UploadFile
+from cloudinary_services import upload_image
 
 async def get_users(db: AsyncSession):
     result = await db.execute(select(User))
@@ -35,9 +38,41 @@ async def remove_user(
 ):
     result = await db.execute(select(User).where(User.id == user_id))
     user = result.scalar_one_or_none()
+    if not user:
+        raise UserNotFoundError(user_id)
     await db.delete(user)
     await db.commit()
     
+async def update_user_image(
+    user_id: int,
+    image_file: UploadFile,
+    db: AsyncSession
+):
+    result = await db.execute(select(User).where(User.id == user_id))
+    user = result.scalar_one_or_none()
+    if not user:
+        raise UserNotFoundError(user_id)
+
+    #   NO NEED TO HANDLE IF USER IS NONE
+
+    response = upload_image(
+        file=image_file,
+        folder="users"
+    )
+
+    user.profile_picture = response['url']
+    user.profile_picture_public_id = response['public_id']
+    
+    try:
+        await db.commit()
+    except IntegrityError:
+        await db.rollback()
+        #   EXCEPTION FOR IMAGE UPLOADING ERRORS NEEDED
+
+    await db.refresh(user)
+    return user
+    
+
